requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2018.03.07=0
certifi=2018.10.15=py37_0
cffi=1.11.5=py37he75722e_1
cudatoolkit=9.0=h13b8566_0
cudnn=7.1.2=cuda9.0_0
cycler=0.10.0=py37_0
dbus=1.13.2=h714fa37_1
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h9c2bf20_1
intel-openmp=2019.1=144
jpeg=9b=h024ee3a_2
kiwisolver=1.0.1=py37hf484d3e_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.35=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
matplotlib=3.0.1=py37h5429711_0
mkl=2018.0.3=1
mkl_fft=1.0.6=py37h7dd41cf_0
mkl_random=1.0.1=py37h4414c95_1
nccl=1.3.5=cuda9.0_0
ncurses=6.1=hf484d3e_0
ninja=1.8.2=py37h6bb024c_1
numpy=1.15.4=py37h1d66e8a_0
numpy-base=1.15.4=py37h81de0dd_0
openssl=1.0.2p=h14c3975_0
pcre=8.42=h439df22_0
pip=18.1=py37_0
pycparser=2.19=py37_0
pyparsing=2.3.0=py37_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.0=h6e4f718_3
python-dateutil=2.7.5=py37_0
pytorch=0.4.1=py37ha74772b_0
pytz=2018.7=py37_0
pyyaml=3.13=py37h14c3975_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
setuptools=40.6.2=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.11.0=py37_1
sqlite=3.25.3=h7b6447c_0
tk=8.6.8=hbc83047_0
tornado=5.1.1=py37h7b6447c_0
wheel=0.32.3=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
